Lesson 1 code complete.  Lesson 1 notes.

+++ project.py
@@ -5,6 +5,19 @@ from sqlalchemy import create_engine, asc
 from sqlalchemy.orm import sessionmaker
 from database_setup import Base, Restaurant, MenuItem
 
+#Imports for web server session management
+from flask import session as login_session
+import random, string
+
+#Imports for OAuth2
+from oauth2client.client import flow_from_clientsecrets
+from oauth2client.client import FlowExchangeError
+import httplib2
+import json
+from flask import make_response
+import requests
+
+CLIENT_ID = json.loads(open('client_secrets.json','r').read())['web']['client_id']
 
 #Connect to Database and create database session
 engine = create_engine('sqlite:///restaurantmenu.db')
@@ -12,6 +25,130 @@ Base.metadata.bind = engine
 
 DBSession = sessionmaker(bind=engine)
 session = DBSession()
+
+#Create a state token to prevent request forgery.
+# Store it in the session for later validation.
+@app.route('/login')
+def showLogin():
+  state=''.join(random.choice(string.ascii_uppercase+string.digits) for x in xrange(32))
+  login_session['state']=state
+  # return "The current session state is %s" % login_session['state']
+  return render_template('login.html', STATE=state)
+
+#Accept Google connection posts for authentication.
+@app.route('/gconnect', methods=['POST'])
+def gconnect():
+  print "Entering gconnect()..."
+  #Validate state token.
+  if request.args.get('state') != login_session['state']:
+    #print "state token check failed..."
+    #print "request.args.get('state'): " + request.args.get('state')
+    #print "login_session['state']: " + login_session['state']
+    response=make_response(json.dumps('Invalid state parameter.'),401)
+    response.headers['Content-Type']='application/json'
+    return response
+  code = request.data
+  print "State token verified."
+  try:
+    #Upgrade the authorization code into a credentials object.
+    oauth_flow = flow_from_clientsecrets('client_secrets.json',scope='')
+    oauth_flow.redirect_uri='postmessage'
+    credentials = oauth_flow.step2_exchange(code)
+    print "Upgraded authorization code into a credentials object."
+  except FlowExchangeError:
+    response = make_response(json.dumps('Failed to upgrade the authorization code.'),401)
+    response.headers['Content-Type']='application/json'
+    return response
+  #Check the access token is valid.
+  print "Checking access token is valid..."
+  access_token = credentials.access_token
+  url=('https://www.googleapis.com/oauth2/v1/tokeninfo?access_token=%s' % access_token)
+  h = httplib2.Http()
+  result = json.loads(h.request(url, 'GET')[1])
+  print "Loaded token info..."
+  #If there was an error in the access token info, abort.
+  if result.get('error') is not None:
+    response = make_response(json.dumps(result.get('error')), 500)
+    response.headers['Content-Type']='application/json'
+    return response
+  print "Token info was good."
+  #Verify that the access token is used for the intended user.
+  gplus_id=credentials.id_token['sub']
+  if result['user_id'] != gplus_id:
+    response=make_response(json.dumps("Token's user ID doesn't match given user ID."),401)
+    response.headers['Content-Type']='application/json'
+    return response
+  #Verify that the access token is valid for this app.
+  if result['issued_to'] != CLIENT_ID:
+    response=make_response(json.dumps("Token's client ID doesn't match app's."),401)
+    print "Token's client ID doesn't match app's."
+    response.headers['Content-Type']='application/json'
+    return response
+  #Check to see if a user is already logged in.
+  stored_credentials=login_session.get('credentials')
+  stored_gplus_id=login_session.get('gplus_id')
+  if stored_credentials is not None and gplus_id==stored_gplus_id:
+    response=make_response(json.dumps("Current user is already connected."),200)
+    response.headers['Content-Type']='application/json'
+    return response
+  #Authentication was successful
+  #Store access token in session for later use.
+  login_session['credentials']=credentials
+  login_session['gplus_id']=gplus_id
+  #Get user info.
+  userinfo_url="https://www.googleapis.com/oauth2/v1/userinfo"
+  params={'access_token':credentials.access_token,'alt':'json'}
+  answer=requests.get(userinfo_url,params=params)
+
+  data=answer.json()
+
+  #Store user info in session.
+  login_session['username']=data['name']
+  login_session['picture']=data['picture']
+  login_session['email']=data['email']
+
+  output=''
+  output+='<h1>Welcome, '
+  output+=login_session['username']
+  output+='!</h1>'
+  output+='<img src="'
+  output+=login_session['picture']
+  output+='" style="width:300px;height:300px;border-radius:150px;-webkit-border-radius:150px;-moz-border-radius:150px;">'
+  flash("you are now logged in as %s" % login_session['username'])
+  print "done!"
+  return output
+
+
+#DISCONNECT - Revoke a current user's token and reset their login_session.
+@app.route('/gdisconnect')
+def gdisconnect():
+  #Only disconnect if a user currently connected.
+  credentials=login_session.get('credentials')
+  if credentials is None:
+    response=make_response(json.dumps('Current user not connected.'),401)
+    response.headers['Content-Type']='application/json'
+    return response
+  #Execute HTTP GET request to revoke current user token.
+  access_token=credentials.access_token
+  url='https://accounts.google.com/o/oauth2/revoke?token=%s' % access_token
+  h = httplib2.Http()
+  result=h.request(url,'GET')[0]
+  #Check result of revoke request:
+  if result['status']=='200':
+    #Was successful, reset user's session:
+    del login_session['credentials']
+    del login_session['gplus_id']
+    del login_session['username']
+    del login_session['picture']
+    del login_session['email']
+    response=make_response(json.dumps('Successfully disconnected.'),200)
+    response.headers['Content-Type']='application/json'
+    return response
+  else:
+    #For some reason token was invalid:
+    response=make_response(json.dumps('Failed to revoke token for current user.'),400)
+    response.headers['Content-Type']='application/json'
+    return response
 
 
 #JSON APIs to view Restaurant Information
@@ -43,6 +180,10 @@ def showRestaurants():
 #Create a new restaurant
 @app.route('/restaurant/new/', methods=['GET','POST'])
 def newRestaurant():
+  #Check if a user is logged in order to add restaurants.
+  if 'username' not in login_session:
+    #Redirect to login if not.
+    return redirect('/login')
   if request.method == 'POST':
       newRestaurant = Restaurant(name = request.form['name'])
       session.add(newRestaurant)
@@ -55,6 +196,10 @@ def newRestaurant():
 #Edit a restaurant
 @app.route('/restaurant/<int:restaurant_id>/edit/', methods = ['GET', 'POST'])
 def editRestaurant(restaurant_id):
+  #Check if a user is logged in order to edit restaurants.
+  if 'username' not in login_session:
+    #Redirect to login if not.
+    return redirect('/login')
   editedRestaurant = session.query(Restaurant).filter_by(id = restaurant_id).one()
   if request.method == 'POST':
       if request.form['name']:
@@ -68,6 +213,10 @@ def editRestaurant(restaurant_id):
 #Delete a restaurant
 @app.route('/restaurant/<int:restaurant_id>/delete/', methods = ['GET','POST'])
 def deleteRestaurant(restaurant_id):
+  #Check if a user is logged in order to delete restaurants.
+  if 'username' not in login_session:
+    #Redirect to login if not.
+    return redirect('/login')
   restaurantToDelete = session.query(Restaurant).filter_by(id = restaurant_id).one()
   if request.method == 'POST':
     session.delete(restaurantToDelete)
@@ -90,6 +239,10 @@ def showMenu(restaurant_id):
 #Create a new menu item
 @app.route('/restaurant/<int:restaurant_id>/menu/new/',methods=['GET','POST'])
 def newMenuItem(restaurant_id):
+  #Check if a user is logged in order to add menu items.
+  if 'username' not in login_session:
+    #Redirect to login if not.
+    return redirect('/login')
   restaurant = session.query(Restaurant).filter_by(id = restaurant_id).one()
   if request.method == 'POST':
       newItem = MenuItem(name = request.form['name'], description = request.form['description'], price = request.form['price'], course = request.form['course'], restaurant_id = restaurant_id)
@@ -103,7 +256,10 @@ def newMenuItem(restaurant_id):
 #Edit a menu item
 @app.route('/restaurant/<int:restaurant_id>/menu/<int:menu_id>/edit', methods=['GET','POST'])
 def editMenuItem(restaurant_id, menu_id):
-
+  #Check if a user is logged in order to edit menu.
+  if 'username' not in login_session:
+    #Redirect to login if not.
+    return redirect('/login')
     editedItem = session.query(MenuItem).filter_by(id = menu_id).one()
     restaurant = session.query(Restaurant).filter_by(id = restaurant_id).one()
     if request.method == 'POST':
@@ -126,6 +282,10 @@ def editMenuItem(restaurant_id, menu_id):
 #Delete a menu item
 @app.route('/restaurant/<int:restaurant_id>/menu/<int:menu_id>/delete', methods = ['GET','POST'])
 def deleteMenuItem(restaurant_id,menu_id):
+  #Check if a user is logged in order to delete menu.
+  if 'username' not in login_session:
+    #Redirect to login if not.
+    return redirect('/login')
     restaurant = session.query(Restaurant).filter_by(id = restaurant_id).one()
     itemToDelete = session.query(MenuItem).filter_by(id = menu_id).one() 
     if request.method == 'POST':
